+++ changes.md
@@ -1,3 +1,7 @@
+ - added command `aggregateintervals`.  [Documentation here](docs/AggregateIntervals.md)
+
+____
+
  - The `count` module no longer counts genotypes by default.  Use option `--genotypes` to do this.
  - Changed sample id column header in several modules.  They are now all "Sample".
  - Added expr function 'hwe'.  Invoke this with `hwe(homref_count, het_count, homvar_count)`